requirements.txt
asgiref==3.7.2
Django==3.2
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.0.0
django-cleanup==6.0.0
Pillow
drf-yasg==1.20.0
pytz==2023.3
sqlparse==0.4.4
PyJWT==1.7.1
psycopg2
psycopg2-binary
gunicorn